# [WIP] Improve performance of List<T>.Add through use of InvariantArray