--- a/src/System.Private.CoreLib/shared/System/Collections/Generic/List.cs
+++ b/src/System.Private.CoreLib/shared/System/Collections/Generic/List.cs
@@ -23,11 +23,11 @@ namespace System.Collections.Generic
     {
         private const int DefaultCapacity = 4;
 
-        private T[] _items; // Do not rename (binary serialization)
+        private InvariantArray<T> _items; // Do not rename (binary serialization)
         private int _size; // Do not rename (binary serialization)
         private int _version; // Do not rename (binary serialization)
 
-        private static readonly T[] s_emptyArray = new T[0];
+        private static readonly InvariantArray<T> s_emptyArray = new InvariantArray<T>(0);
 
         // Constructs a List. The list is initially empty and has a capacity
         // of zero. Upon adding the first element to the list the capacity is
@@ -50,7 +50,7 @@ namespace System.Collections.Generic
             if (capacity == 0)
                 _items = s_emptyArray;
             else
-                _items = new T[capacity];
+                _items = new InvariantArray<T>(capacity);
         }
 
         // Constructs a List, copying the contents of the given collection. The
@@ -71,8 +71,8 @@ namespace System.Collections.Generic
                 }
                 else
                 {
-                    _items = new T[count];
-                    c.CopyTo(_items, 0);
+                    _items = new InvariantArray<T>(count);
+                    c.CopyTo(_items.Array, 0);
                     _size = count;
                 }
             }
@@ -111,10 +111,10 @@ namespace System.Collections.Generic
                 {
                     if (value > 0)
                     {
-                        T[] newItems = new T[value];
+                        InvariantArray<T> newItems = new InvariantArray<T>(value);
                         if (_size > 0)
                         {
-                            Array.Copy(_items, 0, newItems, 0, _size);
+                            _items.SpanNotNull.Slice(0, _size).CopyTo(newItems.SpanNotNull);
                         }
                         _items = newItems;
                     }
@@ -202,7 +202,7 @@ namespace System.Collections.Generic
         public void Add(T item)
         {
             _version++;
-            T[] array = _items;
+            InvariantArray<T> array = _items;
             int size = _size;
             if ((uint)size < (uint)array.Length)
             {
@@ -280,7 +280,7 @@ namespace System.Collections.Generic
             if (_size - index < count)
                 ThrowHelper.ThrowArgumentException(ExceptionResource.Argument_InvalidOffLen);
 
-            return Array.BinarySearch<T>(_items, index, count, item, comparer);
+            return Array.BinarySearch<T>(_items.Array, index, count, item, comparer);
         }
 
         public int BinarySearch(T item)
@@ -300,7 +300,7 @@ namespace System.Collections.Generic
                 _size = 0;
                 if (size > 0)
                 {
-                    Array.Clear(_items, 0, size); // Clear the elements so that the gc can reclaim the references.
+                    _items.SpanNotNull.Slice(0, size).Clear(); // Clear the elements so that the gc can reclaim the references.
                 }
             }
             else
@@ -368,7 +368,7 @@ namespace System.Collections.Generic
             try
             {
                 // Array.Copy will check for NULL.
-                Array.Copy(_items, 0, array, arrayIndex, _size);
+                Array.Copy(_items.Array, 0, array, arrayIndex, _size);
             }
             catch (ArrayTypeMismatchException)
             {
@@ -388,13 +388,13 @@ namespace System.Collections.Generic
             }
 
             // Delegate rest of error checking to Array.Copy.
-            Array.Copy(_items, index, array, arrayIndex, count);
+            Array.Copy(_items.Array, index, array, arrayIndex, count);
         }
 
         public void CopyTo(T[] array, int arrayIndex)
         {
             // Delegate rest of error checking to Array.Copy.
-            Array.Copy(_items, 0, array, arrayIndex, _size);
+            Array.Copy(_items.Array, 0, array, arrayIndex, _size);
         }
 
         // Ensures that the capacity of this list is at least the given minimum
@@ -602,7 +602,7 @@ namespace System.Collections.Generic
             }
 
             List<T> list = new List<T>(count);
-            Array.Copy(_items, index, list._items, 0, count);
+            _items.SpanNotNull.Slice(index, count).CopyTo(list._items.SpanNotNull);
             list._size = count;
             return list;
         }
@@ -617,7 +617,7 @@ namespace System.Collections.Generic
         // search.
         // 
         public int IndexOf(T item)
-            => Array.IndexOf(_items, item, 0, _size);
+         => Array.IndexOf(_items.Array, item, 0, _size);
 
         int IList.IndexOf(object item)
         {
@@ -641,7 +641,7 @@ namespace System.Collections.Generic
         {
             if (index > _size)
                 ThrowHelper.ThrowArgumentOutOfRange_IndexException();
-            return Array.IndexOf(_items, item, index, _size - index);
+            return Array.IndexOf(_items.Array, item, index, _size - index);
         }
 
         // Returns the index of the first occurrence of a given value in a range of
@@ -661,7 +661,7 @@ namespace System.Collections.Generic
             if (count < 0 || index > _size - count)
                 ThrowHelper.ThrowCountArgumentOutOfRange_ArgumentOutOfRange_Count();
 
-            return Array.IndexOf(_items, item, index, count);
+            return Array.IndexOf(_items.Array, item, index, count);
         }
 
         // Inserts an element into this list at a given index. The size of the list
@@ -678,7 +678,7 @@ namespace System.Collections.Generic
             if (_size == _items.Length) EnsureCapacity(_size + 1);
             if (index < _size)
             {
-                Array.Copy(_items, index, _items, index + 1, _size - index);
+                _items.SpanNotNull.Slice(index, _size - index).CopyTo(_items.SpanNotNull.Slice(index + 1));
             }
             _items[index] = item;
             _size++;
@@ -724,20 +724,20 @@ namespace System.Collections.Generic
                     EnsureCapacity(_size + count);
                     if (index < _size)
                     {
-                        Array.Copy(_items, index, _items, index + count, _size - index);
+                        _items.SpanNotNull.Slice(index, _size - index).CopyTo(_items.SpanNotNull.Slice(index + count));
                     }
 
                     // If we're inserting a List into itself, we want to be able to deal with that.
                     if (this == c)
                     {
                         // Copy first part of _items to insert location
-                        Array.Copy(_items, 0, _items, index, index);
+                        _items.SpanNotNull.Slice(0, index).CopyTo(_items.SpanNotNull.Slice(index));
                         // Copy last part of _items back to inserted location
-                        Array.Copy(_items, index + count, _items, index * 2, _size - index);
+                        _items.SpanNotNull.Slice(index + count, _size - index).CopyTo(_items.SpanNotNull.Slice(index * 2));
                     }
                     else
                     {
-                        c.CopyTo(_items, index);
+                        c.CopyTo(_items.Array, index);
                     }
                     _size += count;
                 }
@@ -827,7 +827,7 @@ namespace System.Collections.Generic
                 ThrowHelper.ThrowArgumentOutOfRangeException(ExceptionArgument.count, ExceptionResource.ArgumentOutOfRange_BiggerThanCollection);
             }
 
-            return Array.LastIndexOf(_items, item, index, count);
+            return Array.LastIndexOf(_items.Array, item, index, count);
         }
 
         // Removes the element at the given index. The size of the list is
@@ -882,7 +882,7 @@ namespace System.Collections.Generic
 
             if (RuntimeHelpers.IsReferenceOrContainsReferences<T>())
             {
-                Array.Clear(_items, freeIndex, _size - freeIndex); // Clear the elements so that the gc can reclaim the references.
+                _items.SpanNotNull.Slice(freeIndex, _size - freeIndex).Clear(); // Clear the elements so that the gc can reclaim the references.
             }
 
             int result = _size - freeIndex;
@@ -902,7 +902,7 @@ namespace System.Collections.Generic
             _size--;
             if (index < _size)
             {
-                Array.Copy(_items, index + 1, _items, index, _size - index);
+                _items.SpanNotNull.Slice(index + 1, _size - index).CopyTo(_items.SpanNotNull.Slice(index));
             }
             if (RuntimeHelpers.IsReferenceOrContainsReferences<T>())
             {
@@ -932,13 +932,13 @@ namespace System.Collections.Generic
                 _size -= count;
                 if (index < _size)
                 {
-                    Array.Copy(_items, index + count, _items, index, _size - index);
+                    _items.SpanNotNull.Slice(index + count, _size - index).CopyTo(_items.SpanNotNull.Slice(index));
                 }
 
                 _version++;
                 if (RuntimeHelpers.IsReferenceOrContainsReferences<T>())
                 {
-                    Array.Clear(_items, _size, count);
+                    _items.SpanNotNull.Slice(_size, count).Clear();
                 }
             }
         }
@@ -969,7 +969,7 @@ namespace System.Collections.Generic
 
             if (count > 1)
             {
-                Array.Reverse(_items, index, count);
+                Array.Reverse(_items.Array, index, count);
             }
             _version++;
         }
@@ -1009,7 +1009,7 @@ namespace System.Collections.Generic
 
             if (count > 1)
             {
-                Array.Sort<T>(_items, index, count, comparer);
+                Array.Sort<T>(_items.Array, index, count, comparer);
             }
             _version++;
         }
@@ -1023,7 +1023,7 @@ namespace System.Collections.Generic
 
             if (_size > 1)
             {
-                ArraySortHelper<T>.Sort(_items, 0, _size, comparison);
+                ArraySortHelper<T>.Sort(_items.Array, 0, _size, comparison);
             }
             _version++;
         }
@@ -1034,12 +1034,10 @@ namespace System.Collections.Generic
         {
             if (_size == 0)
             {
-                return s_emptyArray;
+                return s_emptyArray.Array;
             }
 
-            T[] array = new T[_size];
-            Array.Copy(_items, 0, array, 0, _size);
-            return array;
+            return _items.SpanNotNull.Slice(0, _size).ToArray();
         }
 
         // Sets the capacity of this list to the size of the list. This method can