### Description

This is a follow-up to https://github.com/dotnet/coreclr/pull/17386. The primary difference is that it keeps all unsafe code contained within a single type `InvariantArray<T>` and plumbs this type throughout the `List<T>` class. No unsafe code is added to `List<T>`.

This introduces an internal `InvariantArray<T>` type which serves as an invariant wrapper around `T[]` instances. The primary benefit of this is that _writing_ reference types _T_ to the underlying array becomes faster because no type safety checks are necessary. The only check implemented by the runtime is the check that updates the card table.

Some quick benchmarks follow, calling in a loop `List<object>.Add` (with a non-null _object_) and `List<object>.Add` (with a non-null _string_).

|               Method |      Toolchain |     Mean |     Error |    StdDev | Ratio | RatioSD |
|--------------------- |--------------- |---------:|----------:|----------:|------:|--------:|
|    AddObjToListOfObj |     3.0-master | 27.35 us | 0.5356 us | 0.7150 us |  1.00 |    0.00 |
|    AddObjToListOfObj | invariantarray | 24.38 us | 0.4667 us | 0.4365 us |  0.89 |    0.03 |
|                      |                |          |           |           |       |         |
| AddStringToListOfObj |     3.0-master | 37.83 us | 0.7934 us | 0.8818 us |  1.00 |    0.00 |
| AddStringToListOfObj | invariantarray | 24.13 us | 0.2350 us | 0.1963 us |  0.64 |    0.02 |

Adding any object (regardless of type) to a `List<object>` takes the same amount of time in the new branch. This clearly demonstrates that the runtime is skipping all type checks.

Old assembly for `List<T>.Add` (for reference type _T_):

```asm
00007ffc`7409b64d ff4714          inc     dword ptr [rdi+14h]
00007ffc`7409b650 488b4f08        mov     rcx,qword ptr [rdi+8]
00007ffc`7409b654 8b5710          mov     edx,dword ptr [rdi+10h]
00007ffc`7409b657 395108          cmp     dword ptr [rcx+8],edx
00007ffc`7409b65a 7612            jbe     TO_RESIZE (00007ffc`7409b66e)
00007ffc`7409b65c 448d4201        lea     r8d,[rdx+1]
00007ffc`7409b660 44894710        mov     dword ptr [rdi+10h],r8d
00007ffc`7409b664 4c8bc6          mov     r8,rsi
00007ffc`7409b667 e894cc8c5f      call    CoreCLR!JIT_Stelem_Ref (00007ffc`d3968300)
```

New assembly for `List<T>.Add` (for reference type _T_):

```asm
00007ffc`7409d70a ff470c          inc     dword ptr [rdi+0Ch]
00007ffc`7409d70d 488b5710        mov     rdx,qword ptr [rdi+10h]
00007ffc`7409d711 8b4f08          mov     ecx,dword ptr [rdi+8]
00007ffc`7409d714 8b4208          mov     eax,dword ptr [rdx+8]
00007ffc`7409d717 3bc1            cmp     eax,ecx
00007ffc`7409d719 7618            jbe     TO_RESIZE (00007ffc`7409d733)
00007ffc`7409d71b 8d4101          lea     eax,[rcx+1]
00007ffc`7409d71e 894708          mov     dword ptr [rdi+8],eax
00007ffc`7409d721 4863c9          movsxd  rcx,ecx
00007ffc`7409d724 488d4cca10      lea     rcx,[rdx+rcx*8+10h]
00007ffc`7409d729 488bd6          mov     rdx,rsi
00007ffc`7409d72c e8efa38b5f      call    CoreCLR!JIT_WriteBarrier (00007ffc`d3957b20)
```

Of special note above is that the VM has reordered fields within the `List<T>` class (though this seems to have no real impact), and the JIT has chosen to move `_array.Length` into a temporary register _eax_ rather than run the _cmp_ operation directly against the memory address. I don't know why this is, and it increases the codegen ever so slightly, but it doesn't negatively impact perf.

I've confirmed that `List<T>.Add` is also being inlined into its caller, even though the method is not explicitly marked with aggressive inlining.

Benchmarks for non-reference types _T_ are within the realm of noise, as are benchmarks for `List<T>.get_Item`. For example, here's the benchmark for `List<int>.Add(int)`:

|          Method |      Toolchain |      Mean |     Error |    StdDev |    Median | Ratio | RatioSD |
|---------------- |--------------- |----------:|----------:|----------:|----------:|------:|--------:|
| AddIntToIntList |     3.0-master | 14.276 us | 0.2145 us | 0.2006 us | 14.235 us |  1.00 |    0.00 |
| AddIntToIntList | invariantarray | 14.208 us | 0.1654 us | 0.1466 us | 14.178 us |  1.00 |    0.02 |

### Open issues

The current design is marked __WIP__ because it has serialization impact and because it's not unit tested.

The current design replaces a `T[]` field in `List<T>` with an `InvariantArray<T>` field. We can still keep `List<T>` itself serializable by creating a serialization ctor and making the type implement `ISerializable`, but the ability to serialize subclassed types may be impacted.

Another option is to keep the backing field of `List<T>` as a `T[]`, and instead to introduce a factory method `InvariantArray<T>.UnsafeCreate(T[])` which skips type-checking the `T[]` passed in to it. This minimizes the code churn in `List<T>`, and it eliminates the serialization compat break, but it introduces a call to `UnsafeCreate` into `List<T>.Add` and the `List<T>` indexer. (In this PR as originally proposed, the term _unsafe_ never appears anywhere in `List<T>` sources.)