@jkotas 2019-03-29
> `List<object>.Add `(with a non-null string)

How common is to add strings to `List` of objects?