@GrabYourPitchforks 2019-03-29
@jkotas I don't think adding `string` to `List<object>` is all that common. The most common case I think will be where _T_ is the real type of the object (essentially, the _AddObjToListOfObj_ benchmark).

There are also other interesting cases throughout the framework where some components may have a `List<BaseType>` and add instances of derived types to it. Some examples: LINQ uses `List<Expression>`; networking uses `List<HttpContent>`; some other places use `List<Type>`. I don't think these are bottlenecks per se, but the performance difference was interesting to note.