@jkotas 2019-03-29
> There are also other interesting cases throughout the framework where some components may have a `List<BaseType>` and add instances of derived types to it

I agree that there will be cases like this. It may be good enough for those to just improve casting perf. @VSadov mentioned he wants to take a look at that.