@GrabYourPitchforks 2019-03-29
There was also talk of adding escape analysis to the JIT so that it could determine automatically that type checks were unnecessary. I was thinking a bit more about that, and I'm not quite sure how that could work reliably.

Say that write have a method which takes a `List<T>` as a parameter (with reference type _T_). The `List<T>.Add` code is inlined into your method, but at the time your method runs there's no way of knowing where the original `List<T>` instance came from. Perhaps it was constructed via `new List<T>()` (so the backing field is really, honestly a `T[]`). Or perhaps it was constructed through serialization, where the backing field could have a different type `U[]` (because the backing field itself came from the serialized data stream).

In this scenario, we as the implementers of `List<T>` know what our contract is, and we could build explicitly into `List<T>` an assumption that the backing field is always of type `T[]`. But I don't know how escape analysis could ever correctly make that determination on our behalf given that serialization / reflection exists.