@jkotas 2019-03-29
> I was thinking a bit more about that, and I'm not quite sure how that could work reliably.

Yes, the JIT does not have the global view required to make this possible today and it is not going to have it in foreseeable future. This escape analysis pass would likely need to be a static AOT tool.