@VSadov 2019-03-29
JIT/VM could know that no derived types were loaded or instances created and assume the Base is always Base thus skip type checks. The penalty for that is having to invalidate a lot of stuff when Derived enters the system. It is a highly theoretical possibility and might be profitable only in special cases.