@VSadov 2019-03-29
Yes, I am looking at performance of casts. It would be just some form of caching for the cast analysis.