@GrabYourPitchforks 2019-11-07
Is this PR still useful? I'm pessimistic that escape analysis will help with this particular case, and it looks like there's already some investigation re: improving casting performance.

The main scenarios that would benefit would be for components which create `List<T>` instances where `T` is a base type or an interface. Some examples of this are `List<Expression>` (used in LINQ) and `List<JsonNode>` (used in System.Text.Json). But I have no evidence `List<T>.Add` specifically is a bottleneck in these scenarios.